// frontend/src/api/generated/models/ContractHistoryDTO.ts
export interface ContractHistoryDTO {
    /**
     * 
     * @type {number}
     * @memberof ContractHistoryDTO
     */
    id: number;
    /**
     * 
     * @type {number}
     * @memberof ContractHistoryDTO
     */
    contractId: number;
    /**
     * 
     * @type {string}
     * @memberof ContractHistoryDTO
     */
    changeType: string;
    /**
     * 
     * @type {string}
     * @memberof ContractHistoryDTO
     */
    changedAt: string;
    /**
     * 
     * @type {number}
     * @memberof ContractHistoryDTO
     */
    roomId: number;
    /**
     * 
     * @type {number}
     * @memberof ContractHistoryDTO
     */
    amount: number;
    /**
     * 
     * @type {number}
     * @memberof ContractHistoryDTO
     */
    deposit: number;
    /**
     * 
     * @type {boolean}
     * @memberof ContractHistoryDTO
     */
    depositReturned?: boolean | null;
    /**
     * 
     * @type {string}
     * @memberof ContractHistoryDTO
     */
    startDate: string;
    /**
     * 
     * @type {string}
     * @memberof ContractHistoryDTO
     */
    endDate: string;
    /**
     * 
     * @type {string}
     * @memberof ContractHistoryDTO
     */
    terminationDate?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ContractHistoryDTO
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ContractHistoryDTO
     */
    status: string;
    /**
     * 
     * @type {string}
     * @memberof ContractHistoryDTO
     */
    payedTillDayOfMonth: string;
}